+++ R/get_export.R
@@ -3,6 +3,8 @@
 #'   Registry Number.
 #' @param cas_rn (Character) CAS Registry Number in "dash" format, e.g.,
 #'   "50-78-2".
+#' @param verbose (Logical) Should the result be returned as is? Defaults to
+#'   \code{FALSE}.
 #' @details The function performs a sanity check on the provided CAS Registry
 #'   Number and then performs a query. If successful, a character string with
 #'   the available MOL file will be returned.
@@ -14,7 +16,7 @@
 #' }
 #' @importFrom curl curl_fetch_memory handle_setopt new_handle
 #' @export
-get_export <- function(cas_rn) {
+get_export <- function(cas_rn, verbose = FALSE) {
 
   # sanity-check cas_rn
   if (isFALSE(.check_cas(cas_rn))) {
@@ -44,6 +46,15 @@ get_export <- function(cas_rn) {
 
   # decode content
   content <- rawToChar(result$content)
+
+  # check status
+  if(result$status_code != 200L) {
+    if(verbose) {
+      return(content)
+    } else {
+      return(NA_character_)
+    }
+  }
 
   # return content
   content